requirements.txt
aiogram==3.22.0
aiosqlite>=0.21.0
aiomysql>=1.1.2
python-dotenv>=1.1.1